requirements.txt
pandas
psycopg2-binary
SQLAlchemy
pyyaml
python-dotenv
loguru
